### 2019-06-04 #1443 @katyhuff
@ryanstwrt I'm terribly sorry for the delay. I somehow missed this notification! I apologize. I expect to be able to find reviewers today. (related -- have you heard of PyNE? I think you could be an excellent contributor, if you aren't already).

Initially, I know that @kellyrowland would be an exceptional reviewer for this work, if she is available. @kellyrowland :

* @kellyrowland: Dr. Rowland, are you able to and interested in taking on this review?
* This submission is in need of expertise in nuclear engineering as well as familiarity with the systems and languages involved ([tools]).
* Title:
* Summary: FRIDGe: Fast Reactor Input Generator
* Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01443/joss.01443/10.21105.joss.01443.pdf
* Submitting author: @ryanstwrt
* Repository: https://github.com/ryanstwrt/FRIDGe
* Version: v1.0.0
