### 2019-06-04 #1443 @ryanstwrt
@katyhuff No problem! I have heard of PyNE, part of my inspiration for creating FRIDGe was the work everyone with PyNE did to help make NE research more available. I'll look into areas of PyNE where I might be able to contribute, thanks for the info.
