### 2019-06-04 #1443 @katyhuff
Thanks @kellyrowland!

@pshriwise I think you would be an excellent reviewer for this JOSS submission. Some more information about reviewing for JOSS can be found here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

@pshriwise: Dr. Shriwise, are you able to and interested in taking on this review for the Journal of Open Source Software?
This submission is in need of expertise in nuclear engineering as well as familiarity with the systems and languages involved (MCNP, Python, yaml, LaTeX).
Title: FRIDGe: Fast Reactor Input Generator
Article Proof: https://github.com/openjournals/joss-papers/blob/joss.01443/joss.01443/10.21105.joss.01443.pdf
Submitting author: @ryanstwrt
Repository: https://github.com/ryanstwrt/FRIDGe
Version: v1.0.0
