### 2019-06-04 #1443 @katyhuff
Thanks @pshriwise !

Dear both (@kellyrowland @pshriwise ): A key requirement of JOSS submissions is that a submission should be a significant contribution to the available open source software that either enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler).  Minor ‘utility’ packages, including ‘thin’ API clients, are not acceptable.

I want to let you know that I contemplated whether this package is a minor utility, and I determined it to be somewhat on the boundary of these two. It certainly will make research faster and easier for folks who use it. So - as I tend to be generous with this type of "novelty" and "significance" assessment - I am approving it to move forward into the review. However, it's possible that this package overlaps somewhat with the python features of new MCNP Tools python package (first distributed in MCNP 6.2... https://laws.lanl.gov/vhosts/mcnp.lanl.gov/pdf_files/la-ur-17-21779.pdf). I hope you'll consider this question completely, but generously, when reviewing the statement of need.
