### 2019-06-04 #1486 @whedon
[REVIEW]: FRIDGe: Fast Reactor Input Generator
**Submitting author:** @ryanstwrt (<a href="http://orcid.org/0000-0003-4867-6555">Ryan Stewart</a>)
**Repository:** <a href="https://github.com/ryanstwrt/FRIDGe" target ="_blank">https://github.com/ryanstwrt/FRIDGe</a>
**Version:** v1.0.1
