### 2019-06-04 #1486 @katyhuff
@ryanstwrt The first thing I'd like to note is the image in your paper -- it is too large. Please reduce its width.
