### 2019-06-04 3591e2d
Finished updating 'blank' to 'smear'.
