+++ paper/paper.bib
@@ -12,14 +12,14 @@
 	Author = {Lum, E. and others},
 	Publisher = {International Reactor Physics Experiments Evaluation Project},
 	Number = {EBR2-LMFBR-RESR-001},
-	Title = {Evaluation of Run 138B at Experimental Reactor II, A Prototypic Liquid Metal Fast Breeder Reactor},
+	Title = {{Evaluation of Run 138B at Experimental Reactor II, A Prototypic Liquid Metal Fast Breeder Reactor}},
 	Year = {2018}
 }
 
 @misc{GenIV,
-    Publisher = {Gen IV International Forum},
+    Publisher = {Gen {IV} International Forum},
     Year = {2013},
-    Title = {Generation IV Goals},
+    Title = {Generation {IV} Goals},
     howpublished = {https://www.gen-4.org/gif/jcms/c_9502/generation-iv-goals}
 }
 
