### 2019-06-04 #1486 @ryanstwrt
@katyhuff Yup, I noticed this when I first uploaded my paper but I wasn't sure if I should update it before it went to you. After I update my document, does it automatically update the article proof so I can check my changes?
