### 2019-06-05 #1486 @pshriwise
@ryanstwrt during install I got the following error:

`error: package directory 'frdige/driver' does not exist`

Just a typo it seems. Please find a relevant PR in your repo.
