### 2019-06-05 #1486 @ryanstwrt
@pshriwise Thanks for the catch! I've merged the changes into the master branch.
