### 2019-06-05 #1486 @kellyrowland
@ryanstwrt Please see issue #2 in your code repository for a request regarding clearer installation instructions.
