### 2019-06-06 d338b57
Update README.md
Added additional guidance for installation, and updated running FRIDGe.
