### 2019-06-13 #1486 @kellyrowland
@pshriwise are you able to test that the `EBRII_Driver.i` input file generated with the instructions in the README functions correctly as MCNP6 input?
