### 2019-06-13 #1486 @pshriwise
@kellyrowland I had some trouble too and found that it only worked if I imported the module from the repository rather than from the installed location. I think this is because the required data isn't being installed with the package...

@ryanstwrt I saw that you include package data in the `setup.py` but I think the preferred method these days is to use a `MANIFEST.in` file to make sure that data gets included in the package install. More info on that [here](https://python-packaging.readthedocs.io/en/latest/non-code-files.html)

The following `MANIFEST.in` file did the trick for me:

```
include fridge/data/CotN/*.yaml
include fridge/data/assembly/*.yaml
include fridge/data/materials/*.yaml
include fridge/data/core/*.yaml
include fridge/fridge_input_file/*.yaml
include fridge/mcnp_input_files/*.i
```
