### 2019-06-14 b8d1dbe
Included MANIFEST.in file for package data.
