### 2019-06-14 #1486 @ryanstwrt
@pshriwise I've added the `MANIFEST.in` file and left  `include_package_data=True`. I've also removed the `package_data={...}` from the `setup.py` file. I believe this should address the problem @kellyrowland was experiencing. Thank you for the information on the preferred methodology for including package data. Let me know if the problem persists, and I can take some more time to investigate.
