### 2019-06-22 d421175
Updated MCNP reference with DOI.
