+++ paper/paper.bib
@@ -1,12 +1,16 @@
-@article{Goo2008,
-	Author = {Goorley, T.},
-	Journal = {Nuclear Technology},
-	Keywords = {MCNP},
-	Pages = {298--315},
-	Title = {MCNP6 Release Overview},
-	Volume = {180},
-	Year = {2012}
+
+@article{doi:10.13182/NT11-135,
+    author = {T. Goorley},
+    title = {Initial MCNP6 Release Overview},
+    journal = {Nuclear Technology},
+    volume = {180},
+    number = {3},
+    pages = {298-315},
+    year  = {2012},
+    publisher = {Taylor & Francis},
+    doi = {10.13182/NT11-135},
 }
+
 
 @book{Lum2018,
 	Author = {Lum, E. and others},
