### 2019-06-25 976b97e
Updated required Python version for setup and documentation.
