### 2019-07-02 10e0377
Adding the ability to loop over and adjust the assembly or core details.
