### 2019-07-09 #1486 @kellyrowland
@ryanstwrt I don't see the MCNP DOI in the latest article proof - it looks like the `.bib` file was updated with the DOI but the `.md` file still has the previous reference.
