### 2019-07-09 #1486 @kellyrowland
The report by Lum et al. has a DOI at https://doi.org/10.2172/1415120 which should be included as well.
