### 2019-07-11 #1486 @ryanstwrt
@kellyrowland I see the reference mistake for MCNP, and I've adjusted it to reference the updated source. I'll upload it here shortly. As for the report by Lem et al., the DOI you're referencing was a report that was sent to DOE, not the final paper that was included in the IRPhEP handbook. From what I can tell the handbook does not have a DOI.
