### 2019-07-11 e5cfaed
Updated DOI for MCNP in papaer.md
