+++ paper/paper.bib
@@ -1,5 +1,5 @@
 
-@article{doi:10.13182/NT11-135,
+@article{Goo2012,
     author = {T. Goorley},
     title = {Initial MCNP6 Release Overview},
     journal = {Nuclear Technology},
