### 2019-07-19 #1486 @katyhuff
@ryanstwrt thanks for your submission. I'm going through some final checks and a few things popped up.

- [ ] Are there any individuals who have made substantive contributions to this work who are not included in the author list? Some roles in a project warrant co-authorship, and some do not. Really, only the community around the project can decide this. I often consider people who satisfy the roles delineated by this credit project (https://www.casrai.org/credit.html). For example, perhaps Prof. Palmer has served a conceptualilzation or support role. Whether that rises to the level of co-authorship is up to you, but I wanted to bring it up just in case.
- [ ] The citation regarding GenIV goals may be more robust if you include a URL, or if you refer to the technical report. I think the appropriate URL is https://www.gen-4.org/gif/jcms/c_9502/generation-iv-goals . Meanwhile, the appropriate technical report is probably this one... https://www.gen-4.org/gif/jcms/c_60729/technology-roadmap-update-2013 . To include the URL, you might try adding it to your bibfile entry as "url=" or "extra=". For the report, consider:

```bibtex
@techreport{behar_technology_2014,
	type = {Generation {IV} {International} {Forum}},
	title = {Technology {Roadmap} {Update} for {Generation} {IV} {Nuclear} {Energy} {Systems}},
	shorttitle = {{GIF} {Technology} {Roadmap}},
	url = {https://www.gen-4.org/gif/upload/docs/application/pdf/2014-03/gif-tru2014.pdf},
	language = {en},
	number = {January 2014 Update},
	institution = {OECD Nuclear Energy Agency},
	author = {Behar, Christophe},
	month = jan,
	year = {2014},
	note = {https://www.gen-4.org/gif/upload/docs/application/pdf/2014-03/gif-tru2014.pdf},
	pages = {66}
```
- [ ] I think the YAML citation could also be stronger, particularly with the information available at https://yaml.org/spec/1.2/spec.html and here https://yaml.org/spec/cvs/spec.pdf . The authors are copyrighted for each version in this package, so it's reasonable to include them as authors in the citation, and to cite the (very stable) version 1.2.
- [ ] It is appropriate to summarize the state of the field, albeit briefly, just to express clearly how your tool expands on the existing capabilities in the literature. There are python packages with similar capabilities. For example, MCNP has come out with an MCNPTools package giving MCNP a python API (https://mcnp.lanl.gov/pdf_files/la-ur-17-21779.pdf) and PyNE has some, but certainly not all, of these capabilities. I think it's clear your tool is for a particular custom application, but you should make it clear how it goes beyond what exists.
