### 2019-07-19 d3e0683
Updating paper and bib
