+++ paper/paper.bib
@@ -11,23 +11,58 @@
     doi = {10.13182/NT11-135},
 }
 
-
 @book{Lum2018,
 	Author = {Lum, E. and others},
 	Publisher = {International Reactor Physics Experiments Evaluation Project},
 	Number = {EBR2-LMFBR-RESR-001},
 	Title = {{Evaluation of Run 138B at Experimental Reactor {II}, A Prototypic Liquid Metal Fast Breeder Reactor}},
-	Year = {2018}
+	Year = {2018},
 }
 
-@misc{GenIV,
-    Publisher = {Gen {IV} International Forum},
-    Year = {2013},
-    Title = {{Generation IV Goals}},
-    howpublished = {https://www.gen-4.org/gif/jcms/c_9502/generation-iv-goals}
+@techreport{GenIV,
+	type = {Generation {IV} {International} {Forum}},
+	title = {Technology {Roadmap} {Update} for {Generation} {IV} {Nuclear} {Energy} {Systems}},
+	shorttitle = {{GIF} {Technology} {Roadmap}},
+	url = {https://www.gen-4.org/gif/upload/docs/application/pdf/2014-03/gif-tru2014.pdf},
+	language = {en},
+	number = {January 2014 Update},
+	institution = {OECD Nuclear Energy Agency},
+	author = {Behar, Christophe},
+	month = jan,
+	year = {2014},
+	note = {https://www.gen-4.org/gif/upload/docs/application/pdf/2014-03/gif-tru2014.pdf},
+	pages = {66},
 }
 
-@misc{YAML,
-    title = {YAML: YAML Ain't Markup Language},
-    howpublished = {https://yaml.org/}
+@techreport{YAML,
+    title = {{YAML}: {YAML} {Ain't} {Markup} {Language} {Version} 1.2},
+    author = {Ben-Kiki, Orsen and Evans, Clark and dot Net, Ingy,
+    note = {https://yaml.org/},
+    url = {https://yaml.org/spec/cvs/spec.pdf},
+    year = {2009},
+    pages = {84},
+    language = {en},
+    month = oct,
+    edition = {3rd},
+}
+
+@techreport(Solomon2017,
+    title= {The {MCNPTools} {Package}: {Installation} and {Use}},
+    author = {Solomon, Clell J. (CJ) and Bates, Cameron and Kulesza, Joel},
+    year = {2017},
+    month = mar,
+    note = {LA-UR-17-21779},
+    institution - {Los Alamos National Laboratory},
+    pages = {23},
+    language = {en},
+}
+
+@inproceedings{Scopatz2012b,
+    author = {Anthony Scopatz and Paul K. Romano and Paul P.H. Wilson and Kathryn D. Huff},
+    address = {San Diego, CA, USA},
+    booktitle = {Am. Nuc. Soc. Winter Meeting 2012},
+    month = Nov,
+    title = {{PyNE: Python for Nuclear Engineering}},
+    volume = {107},
+    year = {2012}
 }
