### 2019-07-20 #1486 @ryanstwrt
@katyhuff Thanks for the comments. I've updated both the YAML and Gen IV citation to be more specific and provide a better reference for those who are interested. Along with this, I have added a few sentence presenting some work that has already been with MCNP and differentiating FRIDGe from these resources. I did take a look at CRediT and while Dr. Palmer has been instrumental in helping as a grad student, I am not sure if it is appropriate for me to put him on this paper as he has not been involved in the development of it. Hopefully I addressed all of your concerns. If not, please feel free to let me know and I will update accordingly.
