+++ paper/paper.bib
@@ -36,7 +36,7 @@
 
 @techreport{YAML,
     title = {{YAML}: {YAML} {Ain't} {Markup} {Language} {Version} 1.2},
-    author = {Ben-Kiki, Orsen and Evans, Clark and dot Net, Ingy,
+    author = {Ben-Kiki, Orsen and Evans, Clark and dot Net, Ingy},
     note = {https://yaml.org/},
     url = {https://yaml.org/spec/cvs/spec.pdf},
     year = {2009},
