### 2019-07-20 9b5530f
Updating paper and bib
