+++ paper/paper.bib
@@ -46,7 +46,7 @@
     edition = {3rd},
 }
 
-@techreport(Solomon2017,
+@techreport{Solomon2017,
     title= {The {MCNPTools} {Package}: {Installation} and {Use}},
     author = {Solomon, Clell J. (CJ) and Bates, Cameron and Kulesza, Joel},
     year = {2017},
