### 2019-07-20 ab8da66
Updating paper and bib
