+++ paper/paper.bib
@@ -52,7 +52,7 @@
     year = {2017},
     month = mar,
     note = {LA-UR-17-21779},
-    institution - {Los Alamos National Laboratory},
+    institution = {Los Alamos National Laboratory},
     pages = {23},
     language = {en},
 }
