### 2019-07-20 eef764c
Updating paper and bib
