+++ fridge/driver/fridge_driver.py
@@ -14,7 +14,7 @@ import fridge.driver.reactorMaker as reactorMaker
 def main(file_name, **kwargs):
     print('Welcome to FRIDGe, the Fast Reactor Input Deck Generator!')
     global_vars = gb.GlobalVariables()
-    global_vars.read_input_file(file_name, kwargs)
+    global_vars.read_input_file(file_name, **kwargs)
     print(global_vars.input_type)
     if global_vars.input_type == 'Single':
         print('Creating assembly: {}... Please Wait'.format(global_vars.file_name))
