### 2019-08-03 d2b2aca
Added ability to create a material with atom densities.
