+++ fridge/data/materials/FFTF_LiquidNa.yaml
@@ -1,0 +1,7 @@
+Name: FFTF Liquid Sodium
+Elements: [Na]
+Elemental ZAIDs: [11000]
+Isotopic ZAIDs: [11000]]
+Isotopic Weight Percents: [[2.3620E-2]]
+Density: 3.3620E-2
+Linear Coefficient of Expansion: 0.0
