+++ fridge/data/materials/FFTF_Inconel600.yaml
@@ -1,0 +1,7 @@
+Name: FFTF Inconel 600
+Elements: [C, Si, S, Cr, Mn, Fe, Co, Ni, Cu]
+Elemental ZAIDs: [6000, 14000, 16000, 24000, 25000, 26000, 27000, 28000, 29000]
+Isotopic ZAIDs: [[6000], [14000], [16000], [24000], [25000], [26000], [27000], [28000], [29000]]
+Isotopic Weight Percents: [[2.0908E-3], [4.4707E-4], [1.1747E-5], [1.4972E-2], [1.5998E-3], [7.1948E-3], [8.5222E-5], [6.3016E-2], [1.9759E-4]]
+Density: 8.9615E-2
+Linear Coefficient of Expansion: 0.0
