+++ fridge/data/materials/FFTF_IP.yaml
@@ -1,0 +1,9 @@
+Name: FFTF_IP
+Elements: ['U', 'O']
+Elemental ZAIDs: [92000,  8000]
+Isotopic ZAIDs: [[92234, 92235, 92236, 92238],
+                            [8016, 8017, 8018]]
+Isotopic Atom Percents: [[1.2746E-6, 1.6686E-4, 0.0, 2.3007E-2],
+                         [4.6351E-2]]
+Density: 6.9526E-2
+Linear Coefficient of Expansion: 0.0
