+++ fridge/data/materials/FFTF_IF.yaml
@@ -1,0 +1,10 @@
+Name: FFTF Inner Fuel
+Elements: ['U', 'Np', 'Pu', 'Am', 'O']
+Elemental ZAIDs: [92000, 93000, 94000, 95000, 8000]
+Isotopic Atom Percents: {92234: 9.9319E-7, 92235: 1.1417E-4, 92238: 1.5764E-2,
+                         93237: 1.6063E-5,
+                         94238: 3.1181E-6, 94239: 5.1998E-3, 94240: 7.0298E-4, 94241: 6.9284E-5, 94242: 1.2825E-5,
+                         95241: 1.1744E-5,
+                         8016: 4.2690E-2}
+Density: 6.4584E-2
+Linear Coefficient of Expansion: 0.0
