+++ fridge/data/materials/FFTF_PlenumMat.yaml
@@ -1,0 +1,7 @@
+Name: FFTF Plenum Material
+Elements: [B, C, N, Na, Al, Si, P, S, V, Cr, Mn, Fe, Co, Ni, Cu, As, Nb, Mo, Ta]
+Elemental ZAIDs: [5000, 6000, 7000, 11000, 13000, 14000, 15000, 16000, 23000, 24000, 25000, 26000, 27000, 28000, 29000, 33000, 41000, 42000, 73000]
+Isotopic ZAIDs: [[5010, 5011], [6000], [7000], [11000], [13000], [14000], [15000], [16000], [23000], [24000], [25000], [26000], [27000], [28000], [29000], [33000], [41000], [42000], [73000]]
+Isotopic Weight Percents: [[1.2816E-7, 5.1585E-7], [5.7968E-5], [4.9707E-6], [9.6390E-3], [1.2902E-5], [1.8592E-4], [4.4957E-6], [2.1713E-6], [2.7335E-5], [4.6866E-3], [4.4356E-4], [1.5973E-2], [5.9070E-6], [3.2028E-3], [1.0956E-5], [2.7879E-6], [3.7470E-6], [3.6285E-4], [3.8477E-7]]
+Density: 3.4628E-2
+Linear Coefficient of Expansion: 0.0
