+++ fridge/data/materials/FFTF_OF.yaml
@@ -1,0 +1,15 @@
+Name: FFTF_OF
+Elements: ['U', 'Np', 'Pu', 'Am', 'O']
+Elemental ZAIDs: [92000, 93000, 94000, 95000, 8000]
+Isotopic ZAIDs: [[92234, 92235, 92236, 92238], 
+                            [93273], 
+                            [94238, 94239, 94240, 94241, 94242], 
+                            [95241], 
+                            [8016, 8017, 8018]]
+Isotopic Atom Percents: [[1.0327E-6, 1.2127E-4, 0.0, 1.6744E-2],
+                         [1.3313E-5],
+                         [2.6497E-6, 4.2072E-3, 5.6618E-4, 5.7297E-5, 9.2450E-6],
+                         [1.4025E-5],
+                         [4.3372E-2]]
+Density: 6.5108E-2
+Linear Coefficient of Expansion: 0.0
