### 2019-08-03 5e53d93
Updated testing for new atom density materials.
