### 2019-08-04 bdda6fe
Updated testing for new atom density materials. Added Co and Cu elements.
