+++ fridge/test_suite/material_test.py
@@ -282,6 +282,34 @@ def test_material_FFTF_IF():
     for k, v in atomPercentKnown.items():
         assert np.allclose(m.atomPercent[k], v, 5)
 
+def test_material_FFTF_Inconel600():
+    """
+    Test material is the inner fuel from the FFTF Benchmark in the
+    International Handbook of Evaluated Reactor Experiments.
+    """
+    m = materialReader.Material()
+    m.set_material('FFTF_Inconel600')
+    assert m.name == 'FFTF Inconel 600'
+    assert m.elements == ['C', 'Si', 'S', 'Cr', 'Mn', 'Fe', 'Co', 'Ni', 'Cu']
+    assert m.zaids == [6000, 14000, 16000, 24000, 25000, 26000, 27000, 28000, 29000]
+    assert m.isotopicAtomPercents == {6000: 2.0908E-3, 14000: 4.4707E-4, 16000: 1.1747E-5, 24000: 1.4972E-2, 25000: 1.5998E-3,
+                         26000: 7.1948E-3, 27000: 8.5222E-5, 28000: 6.3016E-2, 29000: 1.9759E-4}
+    assert m.enrichmentZaids == []
+    assert m.enrichmentIsotopes == []
+    assert m.enrichmentVector == []
+    assert m.density == 8.9615E-2
+    assert np.allclose(m.atomDensity, 6.4584E-2, 5)
+    atomPercentKnown = {6000: 2.0908E-3,
+                        14028: 4.1230E-4, 14029: 2.0945E-5, 14030: 1.3823E-5,
+                        16032: 1.1158E-5, 16033: 8.8810E-8, 16034: 4.9925E-7, 16036: 1.1747E-9,
+                        24050: 6.5053E-4, 24052: 1.2545E-2, 24053: 1.4225E-3, 24054: 3.5409E-4,
+                        25055: 1.5998E-3,
+                        26054: 4.2054E-4, 26056: 6.6015E-3, 26057: 1.5246E-4, 26058: 2.0289E-5,
+                        27059: 8.5222E-5,
+                        28058: 4.2899E-2, 28060: 1.65246E-2, 28061: 7.1838E-4, 28062: 2.2906E-3, 28064: 5.8290E-4,
+                        29063: 1.3663E-4, 29065: 6.0957E-5}
+    for k, v in atomPercentKnown.items():
+        assert np.allclose(m.atomPercent[k], v)
 
 def test_material_BadWtPer():
     m = materialReader.Material()
