### 2019-08-04 #1486 @katyhuff
@labarba Thank you for the ping.

@pshriwise and @kellyrowland Thank you for your reviews -- we couldn't do this without you.

@ryanstwrt Thank you for your submission and for engaging actively in the review process! I have looked over the paper, double-checked all the DOI links, and have conducted a high-level review of the code itself. Thank you for handling my remaining comments.

Everything looks ship-shape to me. At this point, please double-check the paper yourself, review any lingering details in your code/readme/etc., and then make an archive of the reviewed software in Zenodo/figshare/other service. Please be sure that the DOI metadata (title, authors, etc.) matches this JOSS submission. Once that's complete, please update this thread with the DOI of the archive, and I'll move forward with accepting the submission. Until then, now is your moment for final touchups!
