### 2019-08-05 #1486 @labarba
Before the archive, we will need a new tagged release with all the changes during this review
https://github.com/ryanstwrt/FRIDGe/releases
