### 2019-08-06 2948a19
Minor Update to database.py and material_test.py.
