### 2019-08-06 d201c11
Minor edits to paper.
