### 2019-08-06 574686d
Final edits to paper.
