### 2019-08-06 #1486 @ryanstwrt
I have released a new version of FRIDGe and archived it via Zenodo.
Here is the DOI for the archive: 10.21105/joss.01486
Feel free to let me know if you need anything else.
