### 2019-08-06 #1486 @arfon
@ryanstwrt - please don't use `10.21105/joss.01486` as the DOI for the archive. This is the DOI that we will use for the JOSS paper.

Please use a Zenodo-issued DOI.
