### 2019-08-07 #1486 @katyhuff
Thanks again to @ryanstwrt for your submission, and thanks very much to @pshriwise @kellyrowland for your reviews. I think this is ready to accept @openjournals/joss-eics !
